pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = ""
authors = ["jillvillany <jillian.villany@ibm.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.12.1"
python-dotenv = "^1.0.0"
langchain = "0.0.331"
openai = "0.28.1"
google-search-results = "^2.4.2"
tweepy = "^4.14.0"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
